chore: renamed max input use case file to total token balance use case
